+++ server/service/carves.go
@@ -4,9 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
+	hostctx "github.com/fleetdm/fleet/v4/server/contexts/host"
 	"github.com/fleetdm/fleet/v4/server/fleet"
+	"github.com/google/uuid"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -130,4 +133,102 @@ func (svc *Service) GetBlock(ctx context.Context, carveId, blockId int64) ([]byt
 	}
 
 	return data, nil
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Begin File Carve
+////////////////////////////////////////////////////////////////////////////////
+
+type carveBeginRequest struct {
+	NodeKey    string `json:"node_key"`
+	BlockCount int64  `json:"block_count"`
+	BlockSize  int64  `json:"block_size"`
+	CarveSize  int64  `json:"carve_size"`
+	CarveId    string `json:"carve_id"`
+	RequestId  string `json:"request_id"`
+}
+
+type carveBeginResponse struct {
+	SessionId string `json:"session_id"`
+	Success   bool   `json:"success,omitempty"`
+	Err       error  `json:"error,omitempty"`
+}
+
+func (r carveBeginResponse) error() error { return r.Err }
+
+func carveBeginEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*carveBeginRequest)
+
+	payload := fleet.CarveBeginPayload{
+		BlockCount: req.BlockCount,
+		BlockSize:  req.BlockSize,
+		CarveSize:  req.CarveSize,
+		CarveId:    req.CarveId,
+		RequestId:  req.RequestId,
+	}
+
+	carve, err := svc.CarveBegin(ctx, payload)
+	if err != nil {
+		return carveBeginResponse{Err: err}, nil
+	}
+
+	return carveBeginResponse{SessionId: carve.SessionId, Success: true}, nil
+}
+
+const (
+	maxCarveSize = 8 * 1024 * 1024 * 1024 // 8GB
+	maxBlockSize = 256 * 1024 * 1024      // 256MB
+)
+
+func (svc *Service) CarveBegin(ctx context.Context, payload fleet.CarveBeginPayload) (*fleet.CarveMetadata, error) {
+	// skipauth: Authorization is currently for user endpoints only.
+	svc.authz.SkipAuthorization(ctx)
+
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		return nil, osqueryError{message: "internal error: missing host from request context"}
+	}
+
+	if payload.CarveSize == 0 {
+		return nil, osqueryError{message: "carve_size must be greater than 0"}
+	}
+
+	if payload.BlockSize > maxBlockSize {
+		return nil, osqueryError{message: "block_size exceeds max"}
+	}
+	if payload.CarveSize > maxCarveSize {
+		return nil, osqueryError{message: "carve_size exceeds max"}
+	}
+
+	// The carve should have a total size that fits appropriately into the
+	// number of blocks of the specified size.
+	if payload.CarveSize <= (payload.BlockCount-1)*payload.BlockSize ||
+		payload.CarveSize > payload.BlockCount*payload.BlockSize {
+		return nil, osqueryError{message: "carve_size does not match block_size and block_count"}
+	}
+
+	sessionId, err := uuid.NewRandom()
+	if err != nil {
+		return nil, osqueryError{message: "internal error: generate session ID for carve: " + err.Error()}
+	}
+
+	now := time.Now().UTC()
+	carve := &fleet.CarveMetadata{
+		Name:       fmt.Sprintf("%s-%s-%s", host.Hostname, now.Format(time.RFC3339), payload.RequestId),
+		HostId:     host.ID,
+		BlockCount: payload.BlockCount,
+		BlockSize:  payload.BlockSize,
+		CarveSize:  payload.CarveSize,
+		CarveId:    payload.CarveId,
+		RequestId:  payload.RequestId,
+		SessionId:  sessionId.String(),
+		CreatedAt:  now,
+	}
+
+	carve, err = svc.carveStore.NewCarve(ctx, carve)
+	if err != nil {
+		return nil, osqueryError{message: "internal error: new carve: " + err.Error()}
+	}
+
+	return carve, nil
 }
