+++ server/service/endpoint_carves.go
@@ -8,48 +8,6 @@ import (
 )
 
 ////////////////////////////////////////////////////////////////////////////////
-// Begin File Carve
-////////////////////////////////////////////////////////////////////////////////
-
-type carveBeginRequest struct {
-	NodeKey    string `json:"node_key"`
-	BlockCount int64  `json:"block_count"`
-	BlockSize  int64  `json:"block_size"`
-	CarveSize  int64  `json:"carve_size"`
-	CarveId    string `json:"carve_id"`
-	RequestId  string `json:"request_id"`
-}
-
-type carveBeginResponse struct {
-	SessionId string `json:"session_id"`
-	Success   bool   `json:"success,omitempty"`
-	Err       error  `json:"error,omitempty"`
-}
-
-func (r carveBeginResponse) error() error { return r.Err }
-
-func makeCarveBeginEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(carveBeginRequest)
-
-		payload := fleet.CarveBeginPayload{
-			BlockCount: req.BlockCount,
-			BlockSize:  req.BlockSize,
-			CarveSize:  req.CarveSize,
-			CarveId:    req.CarveId,
-			RequestId:  req.RequestId,
-		}
-
-		carve, err := svc.CarveBegin(ctx, payload)
-		if err != nil {
-			return carveBeginResponse{Err: err}, nil
-		}
-
-		return carveBeginResponse{SessionId: carve.SessionId, Success: true}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Receive Block for File Carve
 ////////////////////////////////////////////////////////////////////////////////
